pyproject.toml
[project]
name = "ja-output-harness"
version = "0.3.5"
description = "Japanese output quality gate for LLM coding agents (Codex CLI/App 等) via MCP finalize server"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "sora", email = "noreply@example.com" }
]
keywords = ["mcp", "japanese", "lint", "quality-gate", "llm", "coding-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Japanese",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
ja-output-tune = "ja_output_harness.tune:main"
ja-output-stats = "ja_output_harness.stats:main"

[project.urls]
Homepage = "https://github.com/sora-bluesky/ja-output-harness"
Repository = "https://github.com/sora-bluesky/ja-output-harness"
Issues = "https://github.com/sora-bluesky/ja-output-harness/issues"

[dependency-groups]
# PEP 735 dependency group. `uv sync` installs this automatically by default,
# so contributors get pytest / ruff / pytest-cov without remembering an
# `--extra dev` flag. Install `pip install -e '.[dev]'` still works because
# the [project.optional-dependencies] mirror below keeps the extras wire
# compatible for pip users who don't read uv docs.
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ja_output_harness"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
